Small optimization to unixTimeConvert
As requested by Dex, something real simple for starters - don't use a
function when not necessary. format() without any variables to insert is
not worthy of a function call, so simply assign using the = operator.

A bunch of similar cases across the whole gamemode so it is fun to see
and learn.