Created own file for IRC code, optimized some...
Arranged /h and /hangup, /p and /pickup commands better (they no longer
contain params[] in the header because they don't use it - the same will
eventually be done to all commands of this sort). Optimized frequent
array access in CMD:hangup.

The IRC code that was moved to vx-irc.pwn has been re-done a bit as
well.